# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Go2TR Back-end Test API",
        "contact": {
            "name": "Meysam Mahmoudi",
            "email": "meysam0011212@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost/GO2TR/public",
            "description": "L5 Swagger OpenApi dynamic host server"
        }
    ],
    "paths": {
        "/api/admin/photo": {
            "get": {
                "tags": [
                    "photo"
                ],
                "summary": "Get all photos",
                "operationId": "/api/admin/photo/index",
                "responses": {
                    "200": {
                        "description": "Returns records",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "photo"
                ],
                "summary": "upload a new photo",
                "operationId": "/api/admin/photo/store",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "description"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "photo",
                                        "type": "string",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "description": {
                                        "description": "photo description",
                                        "type": "string",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "resource created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/photo/{id}": {
            "get": {
                "tags": [
                    "photo"
                ],
                "summary": "Get single photo",
                "operationId": "/api/admin/photo/show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "photo id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns records",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "photo"
                ],
                "summary": "upload a photo",
                "operationId": "/api/admin/photo/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "photo id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file",
                                    "description"
                                ],
                                "properties": {
                                    "file": {
                                        "description": "photo path",
                                        "type": "string",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "description": {
                                        "description": "photo description",
                                        "type": "string",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "resource updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "photo"
                ],
                "summary": "delete a photo",
                "operationId": "/api/admin/photo/destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "photo id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "resource deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/post": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "Get all posts",
                "operationId": "/api/admin/post/index",
                "responses": {
                    "200": {
                        "description": "Returns records",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "post"
                ],
                "summary": "insert a new post",
                "operationId": "/api/admin/post/store",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "active"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "items": {
                                            "type": "string"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "content": {
                                        "type": "string",
                                        "items": {
                                            "type": "string"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "active": {
                                        "type": "integer",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Add new post",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/post/{id}": {
            "get": {
                "tags": [
                    "post"
                ],
                "summary": "Get single post",
                "operationId": "/api/admin/post/show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "photo id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns single pos",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "post"
                ],
                "summary": "update a single post",
                "operationId": "/api/admin/post/update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "active"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "items": {
                                            "type": "string"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "content": {
                                        "type": "string",
                                        "items": {
                                            "type": "string"
                                        },
                                        "collectionFormat": "multi"
                                    },
                                    "active": {
                                        "type": "integer",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Add new post",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "post"
                ],
                "summary": "delete a post",
                "operationId": "/api/admin/post/destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "post id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "resource deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/admin/post/{id}/active": {
            "patch": {
                "tags": [
                    "post"
                ],
                "summary": "active/deactive a post",
                "operationId": "/api/admin/post/activePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "active"
                                ],
                                "properties": {
                                    "active": {
                                        "description": "0,1",
                                        "type": "integer",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "collectionFormat": "multi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Add new post",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/home": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "show all posts",
                "operationId": "home/content",
                "responses": {
                    "200": {
                        "description": "Returns records",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}